+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/impl/BootAccessManagement.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.rse.ccd.lib.boot;
+package uk.gov.hmcts.rse.ccd.lib.impl;
 
 import java.time.Clock;
 import org.springframework.beans.factory.annotation.Value;
@@ -21,7 +21,6 @@ import uk.gov.hmcts.reform.roleassignment.RoleAssignmentApplication;
 import uk.gov.hmcts.reform.roleassignment.config.SecurityConfiguration;
 import uk.gov.hmcts.reform.roleassignment.config.SwaggerConfiguration;
 import uk.gov.hmcts.reform.roleassignment.util.Swagger2SpringBoot;
-import uk.gov.hmcts.rse.ccd.lib.YamlPropertySourceFactory;
 import uk.gov.hmcts.rse.ccd.lib.injected.Common;
 
 @SpringBootConfiguration
