+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/impl/Project.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.rse.ccd.lib;
+package uk.gov.hmcts.rse.ccd.lib.impl;
 
 public enum Project {
   Datastore,
